refactor(unauth_resource_section): create AddEvaluationNoAuth component
refactor(unauth_resource_section): drop unauth functions from AddEvaluation

refactor(unauth_resource_section): split AddEvaluation into Auth and NoUth components
